refactor: delay file util import for fast help
