+++ parakeet_nemo_asr_rocm/transcription/__init__.py
@@ -1,0 +1,5 @@
+"""Transcription subpackage providing CLI helpers."""
+
+from parakeet_nemo_asr_rocm.transcription.cli import cli_transcribe
+
+__all__ = ["cli_transcribe"]
